Moved certificate package to public package to reuse it in other projects
