+++ pkg/kyma/connector/connector.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/tehcyx/kyma-integration/internal/handler"
 
-	"github.com/tehcyx/kyma-integration/internal/certificate"
+	"github.com/tehcyx/kyma-integration/pkg/kyma/certificate"
 	"github.com/tehcyx/kyma-integration/pkg/server"
 )
