+++ pkg/kyma/connector/connector_test.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/tehcyx/kyma-integration/internal/certificate"
+	"github.com/tehcyx/kyma-integration/pkg/kyma/certificate"
 	"github.com/tehcyx/kyma-integration/pkg/server"
 )
