+++ pkg/server/server_test.go
@@ -7,7 +7,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/tehcyx/kyma-integration/internal/certificate"
+	"github.com/tehcyx/kyma-integration/pkg/kyma/certificate"
 	"github.com/tehcyx/kyma-integration/internal/handler"
 )
